# feat(run): Log package name and timing in runScriptInPackageCapturing