--- a/commands/run/package.json
+++ b/commands/run/package.json
@@ -40,6 +40,7 @@
     "@lerna/npm-run-script": "file:../../utils/npm-run-script",
     "@lerna/output": "file:../../utils/output",
     "@lerna/run-parallel-batches": "file:../../utils/run-parallel-batches",
+    "@lerna/timer": "file:../../utils/timer",
     "@lerna/validation-error": "file:../../core/validation-error",
     "p-map": "^1.2.0"
   }