--- a/utils/timer/README.md
+++ b/utils/timer/README.md
@@ -1,0 +1,9 @@
+# `@lerna/timer`
+
+> An internal Lerna tool
+
+## Usage
+
+You probably shouldn't, at least directly.
+
+Install [lerna](https://www.npmjs.com/package/lerna) for access to the `lerna` CLI.