--- a/utils/timer/package.json
+++ b/utils/timer/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "@lerna/timer",
+  "version": "3.0.0",
+  "description": "An internal Lerna tool",
+  "keywords": [
+    "lerna",
+    "utils"
+  ],
+  "homepage": "https://github.com/lerna/lerna",
+  "license": "MIT",
+  "author": {
+    "name": "Kevin Verdieck",
+    "url": "https://github.com/nomcopter"
+  },
+  "files": [
+    "timer.js"
+  ],
+  "main": "timer.js",
+  "engines": {
+    "node": ">= 6.9.0"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/lerna/lerna.git"
+  },
+  "scripts": {
+    "test": "echo \"Run tests from root\" && exit 1"
+  }
+}