--- a/utils/timer/timer.js
+++ b/utils/timer/timer.js
@@ -1,0 +1,13 @@
+"use strict";
+
+module.exports = timer;
+
+function timer() {
+  /* istanbul ignore if */
+  if (process.env.LERNA_INTEGRATION) {
+    return () => 0;
+  }
+
+  const startMillis = Date.now();
+  return () => Date.now() - startMillis;
+}