<!--- Provide a general summary of your changes in the Title above -->
Fixes #1780

## Description
- Adds `@lerna/timer` utility to enable emitting timing information while maintaining test stability.
- Adds log statement in `runScriptInPackageCapturing` to output package name and time taken to run the task.
- Adds timing information to the overall execution of `lerna run`

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->
#1780 

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
`npm link` and then used in local project.

Before:
```
± % lerna run test:lint
lerna notice cli v3.4.3
lerna info Executing command in 29 packages: "yarn run test:lint"
yarn run v1.10.1
$ tslint -c tslint.yml 'types/*'
Done in 1.17s.
yarn run v1.10.1
$ tslint -c tslint.yml 'types/*'
Done in 1.17s.
yarn run v1.10.1
$ tslint -c tslint.yml -p tsconfig.json
Done in 4.68s.
yarn run v1.10.1
$ tslint -c tslint.yml -p tsconfig.json
Done in 4.75s.
yarn run v1.10.1
$ tslint -c tslint.yml -p tsconfig.json
Done in 4.45s.
yarn run v1.10.1
$ tslint -c tslint.yml -p tsconfig.json
Done in 5.69s.
yarn run v1.10.1
$ tslint -c tslint.yml -p tsconfig.json
Done in 5.57s.
yarn run v1.10.1
$ tslint -c tslint.yml -p tsconfig.json
Done in 5.65s.
yarn run v1.10.1
$ tslint -c tslint.yml -p tsconfig.json
Done in 5.68s.
....
```

After:
```
± % lerna run test:lint
lerna notice cli v3.4.3
lerna info Executing command in 29 packages: "yarn run test:lint"
lerna info run Ran npm script 'test:lint' in '@scope/a' in 1.3s:
yarn run v1.10.1
$ tslint -c tslint.yml 'types/*'
Done in 1.10s.
lerna info run Ran npm script 'test:lint' in '@scope/b' in 1.3s:
yarn run v1.10.1
$ tslint -c tslint.yml 'types/*'
Done in 1.10s.
lerna info run Ran npm script 'test:lint' in '@scope/c' in 4.8s:
yarn run v1.10.1
$ tslint -c tslint.yml -p tsconfig.json
Done in 4.61s.
lerna info run Ran npm script 'test:lint' in '@scope/d' in 4.9s:
yarn run v1.10.1
$ tslint -c tslint.yml -p tsconfig.json
Done in 4.66s.
lerna info run Ran npm script 'test:lint' in '@scope/e' in 4.7s:
yarn run v1.10.1
$ tslint -c tslint.yml -p tsconfig.json
Done in 4.41s.
.....
```

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.