@nomcopter 2018-11-20
Thoughts @evocateur?